Tidy up MCE Transceiver STB command definitions. In particular, ensure commands are as succinct as possible and encoding of repeats is correct. The Pronto once sequence is now encoded to contain a full command, not just a command header/part. As a consequence the code for constructing an IrCode from a Pronto command file has been modified. If both once and repeat sequences are present, the repeat sequence does not need to be used.
Note 1: There are some commands which obviously don't match the STB. For example, the Echostar 4720 ENTER.IR command is completely different format/pattern to the other commands for that STB. I've left such commands alone.

Note 2: The only command I found which I could not make any sense of was the Xsat CDTV-410 0.IR command. It seems to be completely borked; perhaps not learned properly?